requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.8.2  # https://github.com/evansd/whitenoise
redis==5.2.1  # https://github.com/redis/redis-py
hiredis==3.1.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==5.0.11  # pyup: < 5.1  # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==65.3.1  # https://github.com/pennersr/django-allauth

# PDF support
ReportLab==4.0.8

# MFA Support
cryptography==41.0.2

# Generative AI Support
google-genai==1.16.1

# Autenticación social
fido2==1.1.1  # Necesario para allauth MFA
PyJWT==2.8.0  # https://github.com/jpadilla/pyjwt
django-crispy-forms==2.3  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-tailwind==1.0.0  # https://github.com/django-crispy-forms/crispy-tailwind
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-ninja==1.3.0  # https://django-ninja.dev/

# Faker
# ------------------------------------------------------------------------------
faker # https://github.com/joke2k/faker
huey[redis]>=2.0.0

# django-sql-explorer
django-sql-explorer

# wagtail
wagtail

# Email processing
django-mailbox>=0.6.9  # https://github.com/coddingtonbear/django-mailbox

# Search
meilisearch==0.31.4  # https://github.com/meilisearch/meilisearch-python
